requirements2.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
dogpile.cache==1.1.5
idna==3.3
jsonschema==3.2.0
pbr==5.9.0
pyrsistent==0.18.1
python-dateutil==2.8.2
requests==2.27.1
rucio-clients==1.28.4
six==1.16.0
stevedore==3.5.0
strictyaml==1.6.1
tabulate==0.8.9
typer==0.4.1
urllib3==1.26.8
